feat: Add phonebook database functionality
This commit adds the necessary code to connect to a MongoDB database and define a schema for the "Person" model. It also includes the transformation of the returned object to include an "id" field instead of "_id" and "__v" fields. This is in preparation for implementing the phonebook database feature.

Refactor the code to use the new database functionality and remove the deprecated "strictQuery" setting.

Closes #3.13, #3.14